requirements.txt
torchvision==0.2.1
gensim==3.4.0
numpy==1.14.5
jieba==0.39